requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.7
django-filter==24.2
djangorestframework==3.15.1
docker==7.1.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
